agent: move common modules
